win-capture: Scooch capture window back if frame nearly yeeted
We wanna seek around and try to align the capture window
over time, such that we're capturing roughly in the middle of the
frame presentation window. Or at least have some healthy margin
before/after, in ms. Not constantly nearly yeeting frames.

If it's possible for the capture window to neatly align in the
middle of the frame presentation window, this should do it within
~5 frames (assuming game FPS and capture FPS are 1:1 identical
FPS rates.)

Otherwise these small adjustments could effectively increase
capture rate, particularly for actually high-refresh-rate games
relative to the capture rate (say, 2:1, 3:1 game to capture rate?),
so this needs thinking through, but for now, just prototyping.